pydantic: split base and charmcraft config models
Common models moved to basic.py

Charmcraft Config model moved to charmcraft.py and renamed
to CharmcraftConfig to avoid confuse with juju's config.
