+++ charmcraft/bases.py
@@ -18,7 +18,7 @@
 
 from typing import Tuple, Union
 
-from charmcraft.models.config import Base
+from charmcraft.models.charmcraft import Base
 from charmcraft.utils import get_host_architecture, get_os_platform
