+++ charmcraft/models/charmcraft.py
@@ -15,12 +15,12 @@
 # For further info, check https://github.com/canonical/charmcraft
 
 """Charmcraft configuration pydantic model."""
-
 import datetime
 import pathlib
 import os
 import re
 import keyword
+
 from typing import Optional, List, Dict, Any
 
 import pydantic
@@ -29,77 +29,7 @@ from craft_cli import CraftError
 from charmcraft.parts import process_part_config
 from charmcraft.utils import get_host_architecture
 from charmcraft.format import format_pydantic_errors
-
-
-class ModelConfigDefaults(
-    pydantic.BaseModel, extra=pydantic.Extra.forbid, frozen=True, validate_all=True
-):
-    """Define Charmcraft's defaults for the BaseModel configuration."""
-
-
-class CustomStrictStr(pydantic.StrictStr):
-    """Generic class to create custom strict strings validated by pydantic."""
-
-    @classmethod
-    def __get_validators__(cls):
-        """Yield the relevant validators."""
-        yield from super().__get_validators__()
-        yield cls.custom_validate
-
-
-class RelativePath(CustomStrictStr):
-    """Constrained string which must be a relative path."""
-
-    @classmethod
-    def custom_validate(cls, value: str) -> str:
-        """Validate relative path.
-
-        Check if it's an absolute path using POSIX's '/' (not os.path.sep, as the charm's
-        config is independent of the platform where charmcraft is running.
-        """
-        if not value:
-            raise ValueError(f"{value!r} must be a valid relative path (cannot be empty)")
-
-        if value[0] == "/":
-            raise ValueError(f"{value!r} must be a valid relative path (cannot start with '/')")
-
-        return value
-
-
-class AttributeName(CustomStrictStr):
-    """Constrained string that must match the name of an attribute from linters.CHECKERS."""
-
-    @classmethod
-    def custom_validate(cls, value: str) -> str:
-        """Validate attribute name."""
-        from charmcraft import linters  # import here to avoid cyclic imports
-
-        valid_names = [
-            checker.name
-            for checker in linters.CHECKERS
-            if checker.check_type == linters.CheckType.attribute
-        ]
-        if value not in valid_names:
-            raise ValueError(f"Bad attribute name {value!r}")
-        return value
-
-
-class LinterName(CustomStrictStr):
-    """Constrained string that must match the name of a linter from linters.CHECKERS."""
-
-    @classmethod
-    def custom_validate(cls, value: str) -> str:
-        """Validate attribute name."""
-        from charmcraft import linters  # import here to avoid cyclic imports
-
-        valid_names = [
-            checker.name
-            for checker in linters.CHECKERS
-            if checker.check_type == linters.CheckType.lint
-        ]
-        if value not in valid_names:
-            raise ValueError(f"Bad lint name {value!r}")
-        return value
+from charmcraft.models.basic import ModelConfigDefaults, AttributeName, LinterName
 
 
 class CharmhubConfig(
@@ -157,7 +87,11 @@ class AnalysisConfig(ModelConfigDefaults, allow_population_by_field_name=True):
     ignore: Ignore = Ignore()
 
 
-class Config(ModelConfigDefaults, validate_all=False):
+class CharmcraftConfig(
+    ModelConfigDefaults,
+    validate_all=False,
+    alias_generator=lambda s: s.replace("_", "-"),
+):
     """Definition of charmcraft.yaml configuration."""
 
     # this needs to go before 'parts', as it used by the validator
