+++ tests/test_config.py
@@ -25,7 +25,7 @@ import pytest
 from craft_cli import CraftError
 
 from charmcraft import linters
-from charmcraft.models.config import Base, BasesConfiguration, CharmhubConfig
+from charmcraft.models.charmcraft import Base, BasesConfiguration, CharmhubConfig
 from charmcraft.config import load
 from charmcraft.utils import get_host_architecture
 
@@ -81,7 +81,7 @@ def test_load_managed_mode_directory(monkeypatch, tmp_path):
     monkeypatch.setenv("CHARMCRAFT_MANAGED_MODE", "1")
 
     # Patch out Config (and Project) to prevent directory validation checks.
-    with patch("charmcraft.config.Config"):
+    with patch("charmcraft.config.CharmcraftConfig"):
         with patch("charmcraft.config.Project") as mock_project:
             with patch("charmcraft.config.load_yaml"):
                 load(None)
