+++ tests/test_manifest.py
@@ -23,7 +23,7 @@ import yaml
 from craft_cli import CraftError
 
 from charmcraft import __version__, linters
-from charmcraft.models.config import Base, BasesConfiguration
+from charmcraft.models.charmcraft import Base, BasesConfiguration
 from charmcraft.metafiles.manifest import create_manifest
 from charmcraft.const import IMAGE_INFO_ENV_VAR
 from charmcraft.utils import OSPlatform
